+++ general/async/go-client/go.mod
@@ -1,6 +1,6 @@
 module github.com/dubbogo/dubbogo-samples
 
 require (
-	github.com/apache/dubbo-go v0.1.2-0.20191115170606-d82e97f3
+	github.com/apache/dubbo-go v0.1.2-0.20191219122427-56cb00a83ba8
 	github.com/apache/dubbo-go-hessian2 v1.2.5-0.20191029001541-894e45c9aaaa
 )
