+++ general/async/go-client/profiles/dev/client.yml
@@ -7,12 +7,6 @@ request_timeout : "3s"
 # connect timeout
 connect_timeout : "3s"
 
-config_center:
-  protocol: apollo
-  address: 127.0.0.1:8080
-  app_id: testApplication_yang
-  cluster: dev
-
 # application config
 application:
   organization : "ikurento.com"
